+++ pytsite/auth/__init__.py
@@ -3,7 +3,7 @@
 # Public API
 from . import _error as error, _model as model, _driver as driver, _widget as widget
 from ._api import get_current_user, get_user_statuses, get_user, create_user, get_role, get_sign_in_form, \
-    register_auth_driver, user_nickname_rule, sign_in, get_auth_driver, create_role, get_sign_in_url, get_sign_out_url,\
+    register_auth_driver, user_nickname_rule, sign_in, get_auth_driver, create_role, get_sign_in_url, get_sign_out_url, \
     verify_password, hash_password, sign_out, get_access_token_info, switch_user, get_anonymous_user, \
     get_system_user, get_users, get_storage_driver, count_users, count_roles, get_first_admin_user, get_roles, \
     get_user_modify_form, base_path, switch_user_to_system, switch_user_to_anonymous, restore_user, \
@@ -49,11 +49,13 @@ def __init():
 
     # Routes
     bp = base_path()
-    router.add_rule(bp + '/sign-in/<driver>', 'pytsite.auth@sign_in')
-    router.add_rule(bp + '/sign-in/<driver>/post', 'pytsite.auth@sign_in_submit', method='POST')
-    router.add_rule(bp + '/sign-out/<driver>', 'pytsite.auth@sign_out')
-    router.add_rule(bp + '/profile/<nickname>', 'pytsite.auth@profile_view')
-    router.add_rule(bp + '/profile/<nickname>/edit', 'pytsite.auth@profile_edit', filters=['pytsite.auth@f_authorize'])
+    router.handle(bp + '/sign-in/<driver>', 'pytsite.auth@sign_in', 'pytsite.auth@sign_in')
+    router.handle(bp + '/sign-in/<driver>/post', 'pytsite.auth@sign_in_submit', 'pytsite.auth@sign_in_submit',
+                  methods='POST')
+    router.handle(bp + '/sign-out/<driver>', 'pytsite.auth@sign_out', 'pytsite.auth@sign_out')
+    router.handle(bp + '/profile/<nickname>', 'pytsite.auth@profile_view', 'pytsite.auth@profile_view')
+    router.handle(bp + '/profile/<nickname>/edit', 'pytsite.auth@profile_edit', 'pytsite.auth@profile_edit',
+                  filters='pytsite.auth@f_authorize')
 
     # Template engine globals
     tpl.register_global('auth_current_user', get_current_user)