+++ pytsite/file_storage_odm/__init__.py
@@ -19,8 +19,8 @@ def _init():
     # Event handlers
     events.listen('pytsite.update', _eh.update)
 
-    router.add_rule('/image/resize/<int:width>/<int:height>/<p1>/<p2>/<filename>', 'pytsite.file_storage_odm@image',
-                    defaults={'width': 0, 'height': 0})
+    router.handle('/image/resize/<int:width>/<int:height>/<p1>/<p2>/<filename>', 'pytsite.file_storage_odm@image',
+                    'pytsite.file_storage_odm@image', defaults={'width': 0, 'height': 0})
 
 
 _init()