+++ pytsite/router/__init__.py
@@ -1,6 +1,6 @@
 """PytSite Router.
 """
-from ._api import add_rule, add_path_alias, base_path, base_url, call_ep, current_path, current_url, dispatch, ep_path, \
+from ._api import handle, add_path_alias, base_path, base_url, call_ep, current_path, current_url, dispatch, ep_path, \
     ep_url, is_base_url, is_ep_callable, get_no_cache, set_no_cache, remove_path_alias, resolve_ep_callable, scheme, \
     server_name, url, session, request, set_request, get_session_store, on_pre_dispatch, on_dispatch, on_response, \
     on_exception