+++ pytsite/routing/_rules_map.py
@@ -1,7 +1,7 @@
-"""PytSite Routing Rule Map.
+"""PytSite Routing Rules Map
 """
 import re as _re
-from typing import Dict as _Dict, Hashable as _Hashable
+from typing import Dict as _Dict
 from . import _rule, _error
 
 __author__ = 'Alexander Shepetko'
@@ -13,38 +13,40 @@ _rule_arg_param_re = _re.compile('\w+:')
 
 
 class RulesMap:
+    """Rules Map
+    """
+
     def __init__(self):
-        self._rules_by_method_path = {}  # type: _Dict[_Hashable, _rule.Rule]
-        self._rules_by_name = {}  # type: _Dict[_Hashable, _rule.Rule]
+        """Init.
+        """
+        self._rules = {}  # type: _Dict[str, _rule.Rule]
 
     def add(self, rule: _rule.Rule):
-        rule_k = (rule.method if rule.method != '*' else 'ANY', rule.path)
+        """Add a rule.
+        """
+        # Check if the rule with the same name is already exist
+        if rule.name in self._rules:
+            raise _error.RuleExists("Rule with name '{}' is already added".format(rule.name))
 
-        # Check if the rule with same pattern but for all methods has been added already
-        if rule_k[0] != 'ANY' and ('ANY', rule_k) in self._rules_by_method_path:
-            raise _error.RuleExists("Rule '{}' is already added".format(rule_k))
-
-        if rule_k in self._rules_by_method_path:
-            raise _error.RuleExists("Rule '{}' is already added".format(rule_k))
-
-        self._rules_by_method_path[rule_k] = rule
-        self._rules_by_name[rule.name] = rule
+        # Add a rule
+        self._rules[rule.name] = rule
 
     def get(self, name: str):
-        """Get rule by name.
+        """Get rule by name
         """
         try:
-            return self._rules_by_name[name]
+            return self._rules[name]
         except KeyError:
             raise _error.RuleNotFound("Rule with name '{}' is not found".format(name))
 
     def match(self, path: str, method: str = 'GET') -> _rule.Rule:
-        if not path.startswith('/'):
-            path = '/' + path
+        """Match rule against a path.
+        """
+        method = method.upper()
 
-        for rule in self._rules_by_name.values():
+        for rule in self._rules.values():
             m = rule.regex.match(path)
-            if not m or rule.method not in (method.upper(), '*'):
+            if not m or method not in rule.methods:
                 continue
 
             # Fill rule's arguments
@@ -53,11 +55,12 @@ class RulesMap:
 
             return rule
 
-        raise _error.RuleNotFound("No rules match path '{}'".format(path))
+        raise _error.RuleNotFound("No rules match the path '{}' against method '{}'".format(path, method))
 
     def path(self, name: str, args: dict = None) -> str:
         """Build path for a rule.
         """
+
         def repl(match):
             nonlocal rule